requirements.txt
torch
torchvision
osmnx
geopandas
rasterio
numpy
Pillow
opencv-python
